Create and share 1 HTTP client.
The client is created when the EXI device starts up, and handles to it
are provided to the inner modules (UserManager, GameReporter, etc). This
allows sharing one connection pool across all clients.

This also renames `SlippiGameReporter` to `GameReporter` to keep
consistency with Jukebox and UserManager naming.

Minor additional cleanup and formatting included.