Adds back support for accessing game memory offset.
This data is only accessible after both the EXI device
and the memory subsystem have initialized, so there's now a new
callback that is notified after initialization of both has completed.

This callback is also where the Discord handler is launched; there may
still be ongoing work here regarding preferences/launching/etc, so WIP.

This also corrects the log name to match the one in the Ishiiruka repo.