Initial port of SlippiDirectCodes to Rust (#18)
This moves the direct codes parsing/loading/vending from C++ to Rust.
Where applicable, certain things that weren't in use in the C++ version
- but may be useful in the future - were also ported but commented out.

The only significant change here is that this changes the `lastPlayed`
field in codes files to be a unix timestamp rather than a datetimestamp
that requires a full parser. There's fallback code in the parser to
handle the older format, though; subsequent writes will just write unix
timestamps to the file instead. This entire bit should be transparent to
users and is purely a convenience thing.

This also makes a change to the public interface of the Rust lib: rather
than passing over a fully qualified `user.json` path, we now just pass
the folder where it's found and then build paths on the Rust side -
since we need to do this for direct code file paths anyway.

* Add an alias for mainline keys.

It looks like the mainline build might use different cased keys, so
adding an alias here so they all centralize and the lib can work across
builds.

* Tweak cbindgen to guard itself a bit more, alter an FFI enum name so that c/cpp do not complain incessantly about overrides

* sort in descending order for lastplayed

---------

Co-authored-by: Nikhil Narayana <nikhil.narayana@live.com>