Add in retry logic for rank API calls.
This migrates the constant background thread setup to one where we only
spawn the background thread when we need to request the rank (i.e on
the CSS). This is ultimately cleaner and makes retry logic easier to
implement: before this change, leaving and re-entering the CSS screen
would likely have hung the client since the background thread would be
stuck in network request sleep timeouts and not be able to pick up the
`Message::FetchRank` event.

(Fizzi should probably tweak the retry timeout per his needs)