fetch match result directly to ensure correctness
before it would have been possible for some abandoned matches to get reported while we were in a game and the rating change would not have shown the impact of the last match but rather the impact of all matches combined. This could make it look like someone lost points when they won, for example.

This new method avoids all of that.